[00071] Move ActivityHeatmap and QRCode tests to widget-local test projects
- Create src/widgets/Ivy.Widgets.ActivityHeatmap/Tests/ with dedicated xunit project
- Create src/widgets/Ivy.Widgets.QRCode/Tests/ with dedicated xunit project
- Update namespaces from Ivy.Test to widget-specific namespaces
- Remove Ivy.Widgets.ActivityHeatmap and Ivy.Widgets.QRCode project references from Ivy.Test.csproj
- Delete ActivityHeatmapTests.cs and QRCodeTests.cs from Ivy.Test
